% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{print_basic_eda}
\alias{print_basic_eda}
\title{Performs basic exploratory data analysis}
\usage{
print_basic_eda(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
a char vector
}
\description{
Performs basic exploratory data analysis
}
